test: update integration-tests
